refactor: shorten route (#159)
