--- src/fsd_entities/ocaasion/api/index.ts
@@ -1,2 +1,0 @@
-export * from './get';
-export * from './post';
