--- src/fsd_entities/ocaasion/model/index.ts
@@ -1,1 +1,0 @@
-export * from './useCeremonyData';
