+++ src/fsd_entities/ocaasion/ui/OccasionApprovalButton.tsx
@@ -3,10 +3,10 @@ const buttonColor = {
   GRAY: 'bg-[#B4B1B1]',
 };
 
-export const OccasionApprovalButton = ({ color, onClick, text }: OccasionApprovalButtonProps) => {
+export const OccasionApprovalButton = ({ color, onClick, text }: Occasion.OccasionApprovalButtonProps) => {
   return (
     <button
-      className={` whitespace-nowrap flex justify-center items-center rounded-lg px-10 py-1 text-lg font-semibold md:px-16 md:text-xl w-1/4 h-[55px] ${buttonColor[color]}`}
+      className={`flex h-[55px] w-1/4 items-center justify-center whitespace-nowrap rounded-lg px-10 py-1 text-lg font-semibold md:px-16 md:text-xl ${buttonColor[color]}`}
       onClick={onClick}
     >
       {text}
