+++ src/fsd_widgets/occasion/index.ts
@@ -1,0 +1,3 @@
+export { OccasionRequestManagement } from './ui/OccasionRequestManagement';
+export { OccasionList } from './ui/OccasionList';
+export { OccasionDetailPage } from './ui/OccasionDetailPage';
