--- src/fsd_widgets/occasion/ui/index.ts
@@ -1,3 +1,0 @@
-export { OccasionRequestManagement } from './OccasionRequestManagement';
-export { OccasionList } from './OccasionList';
-export { OccasionDetailPage } from './OccasionDetailPage';
